Changed to use unit vectors for movement.

+++ Core/Coordinate.cs
@@ -4,8 +4,8 @@ namespace MarsRoverKata.Core
 {
     public class Coordinate
     {
-        public Int32 X { get; set; }
-        public Int32 Y { get; set; }
+        public Double X { get; set; }
+        public Double Y { get; set; }
 
         public Coordinate(Int32 x, Int32 y)
         {
